Refactor Catalog Page
