+++ src/pages/CatalogPage/CatalogPage.module.css
@@ -203,7 +203,7 @@
     display: flex;
     align-items: center;
     justify-content: flex-start;
-    padding: 10px;
+    padding: 10px 20px;
   }
 
   .inputArea {
